requirements.txt
django>=4.0
stripe>=5.0
python-dotenv
django-csp==3.8